build: move from hook to lib
