+++ lib/hash-change.js
@@ -6,7 +6,7 @@ let currentHash
 
 export default fn => {
   useEffect(() => {
-    const hasChange = () => {
+    const hashChange = () => {
       const hash = getHash()
       if (hash !== currentHash) {
         currentHash = hash
@@ -14,7 +14,7 @@ export default fn => {
       }
     }
 
-    window.document.addEventListener('scrollStart', hasChange)
-    return () => window.document.removeEventListener('scrollStart', hasChange)
+    window.document.addEventListener('scrollStart', hashChange)
+    return () => window.document.removeEventListener('scrollStart', hashChange)
   }, [])
 }
